requirements.txt
attrs==25.4.0
aws-cdk-lib==2.232.1
aws-cdk.asset-awscli-v1==2.2.242
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==48.20.0
aws-lambda-powertools==2.43.0
boto3==1.42.16
botocore==1.42.16
cattrs==25.3.0
constructs==10.4.4
exceptiongroup==1.3.1
importlib_resources==6.5.2
jmespath==1.0.1
jsii==1.122.0
pydantic==2.5.3
publication==0.0.3
python-dateutil==2.9.0.post0
s3transfer==0.16.0
six==1.17.0
typeguard==2.13.3
typing_extensions==4.15.0
urllib3==2.6.2

